pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytelegrambotapi>=4.27.0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
    "sqlalchemy>=2.0.41",
]
